workqueue: Add try_queue_work_on() to suggest a CPU
Using queue_work_on() requires that the caller ensure that the specified
CPU stays online.  However, there are cases where performance is best
if a given CPU is used in the common case, but where the occasional
migration to some other CPU is just fine as well.  Therefore, provide
a try_queue_work_on() for which the specified CPU is only a suggestion.

Cc: Tejun Heo <tj@kernel.org>
Cc: Lai Jiangshan <jiangshanlai@gmail.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>